exam-cheating-detection/requirements.txt
# opencv-python>=4.5.0
# facenet-pytorch>=2.5.0       # Replaces mtcnn
# mediapipe>=0.8.0
# torch>=1.7.0
# torchvision>=0.8.0
# flask>=2.0.0
# pyyaml>=5.0.0
# numpy>=1.20.0
# ultralytics>=8.0.0
# psutil>=5.8.0
# mss>=6.1.0

# gTTS>=2.3.1
# pygame>=2.1.2
# playsound>=1.3.0

# # Audio processing
# pyaudio>=0.2.13
# webrtcvad>=2.0.10
# openai-whisper>=20231106
# torchaudio>=2.1.0

# Core Computer Vision
opencv-python>=4.5.0
facenet-pytorch>=2.5.0
mediapipe==0.10.9
ultralytics>=8.0.0

# Machine Learning Framework
torch>=1.7.0
torchvision>=0.8.0
torchaudio>=2.1.0

# Audio Processing
pyaudio>=0.2.13
webrtcvad-wheels>=2.0.10
openai-whisper>=20231106

# Screen Capture
mss>=6.1.0

# Alert System
gTTS>=2.3.1
pygame>=2.1.2


# Reporting System
pdfkit>=1.0.0
jinja2>=3.0.0
matplotlib>=3.5.0
Pillow>=9.0.0  # Required for matplotlib image saving

# System Utilities
psutil>=5.8.0
pyyaml>=5.0.0
numpy>=1.20.0
flask>=2.0.0

# Platform-Specific Dependencies
pywin32>=300; sys_platform == 'win32'
dbus-python>=1.3.2; sys_platform == 'linux'
pygobject>=3.42.2; sys_platform == 'linux'